requirements.txt
absl-py==0.7.1
aniso8601==6.0.0
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
aws-xray-sdk==0.95
bcrypt==3.1.6
boto==2.49.0
boto3==1.7.34
botocore==1.10.84
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
coverage==4.5.3
cryptography==2.7
docker==4.0.1
docutils==0.14
ecdsa==0.13.2
Flask==0.12.2
Flask-Bcrypt==0.7.1
flask-restplus==0.10.1
future==0.17.1
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsondiff==1.1.1
jsonpickle==1.2
jsonschema==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Keras == 2.1.3
Markdown==3.1.1
MarkupSafe==1.1.1
mock==3.0.5
moto==1.3.7
numpy==1.16.4
Pillow==6.0.0
protobuf==3.8.0
pyaml==19.4.1
pycparser==2.19
pycryptodome==3.8.2
pymongo==3.7.1
pyrsistent==0.15.2
python-dateutil==2.8.0
python-jose==2.0.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
responses==0.10.6
s3transfer==0.1.13
schedule==0.6.0
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.14.1
wrapt==1.11.1
xmltodict==0.12.0
kneed==0.3.1
factor_analyzer>=0.3.1
wordcloud>=1.5.0
pandas>=0.23.4
xgboost>=0.90
nltk>=3.4.3
scikit_learn>=0.21.3
